issues fixes

+++ src/main/java/com/sushma/olxlogin/Repository/OlxLoginRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import com.sushma.olxlogin.entity.UserEntity;
 
 @Repository
-public interface OlxLoginRepository extends JpaRepository<UserEntity, Integer>{
+public interface OlxLoginRepository extends JpaRepository<UserEntity, Long>{
 	
 	Optional<UserEntity> findByUserName(String userName);
